Support binary CRCs
Some devices (incorrectly) use binary CRCs. This needs to be taken into
account both when receiving messages and sending responses.

This is currently a per-server option, even though the code does its
best to detect the CRC style used by the client. This is because the
server has to know which CRC style to use even for NAKs sent as a result
of a failure in parsing an incoming message. It may be desirable to
adjust the error handling path to remember the CRC style of the incoming
message instead, so that there does not need to be a config setting.

This patch also includes a couple of bugfixes for the response message
format. The "L" prefix for the "line" was duplicated because it was
both present in the format string and in the "line" property, and the
receiver ended up in the message as "RNone" instead of the desired "R0"
because it is initialized to None.